requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
isodate
pymongo[srv] >= 4.0